+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * more expected drawing when shape has opacity, stroke and shadow
     * `scale` now affect to `shadowOffset`
     * new `perfectDrawEnabled` property for shape. See [http://konvajs.github.io/docs/performance/Disable_Perfect_Draw.html](http://konvajs.github.io/docs/performance/Disable_Perfect_Draw.html)
+    * new `shadowForStrokeEnabled` property for shape. See [http://konvajs.github.io/docs/performance/All_Performance_Tips.html](http://konvajs.github.io/docs/performance/All_Performance_Tips.html)
 
 ## 0.8.0 2015-02-04